backend/requirements.txt
fastapi>=0.68.0
websockets>=10.0
openai>=1.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
